# Improve allocations while doing index queries