--- a/pkg/util/yolo.go
+++ b/pkg/util/yolo.go
@@ -5,3 +5,7 @@ import "unsafe"
 func YoloBuf(s string) []byte {
 	return *((*[]byte)(unsafe.Pointer(&s)))
 }
+
+func YoloString(buf []byte) string {
+	return *((*string)(unsafe.Pointer(&buf)))
+}