This contains 3 optimizations.

1 - Reuse the iterator when iterating through batches
2 - Avoid the string copy when parsing the data from the index.
3 - Use an intermediary `IndexQueryEntry` that does not contains the tableName and hashValue strings for lookupEntriesByQueries that doesn't require it. 

Hard to get a non noisy benchmark but that's how it looks like before and after:

```
benchmark                                   old ns/op     new ns/op     delta
Benchmark_GetRefsChunkWithManyChunks-16     73513148      72606738      -1.23%

benchmark                                   old allocs     new allocs     delta
Benchmark_GetRefsChunkWithManyChunks-16     741487         641464         -13.49%

benchmark                                   old bytes     new bytes     delta
Benchmark_GetRefsChunkWithManyChunks-16     78208523      64817304      -17.12%
```

Performance is unchanged but allocation is reduced.

Why I'm doing this because I found out that we do allocate a lot in size and object.

![image](https://user-images.githubusercontent.com/1053421/99701657-b82c3a80-2a62-11eb-9ed6-8879944cf9d5.png)