@cyriltovena 2020-11-27
> > This contains 3 optimizations.
> 
> I would have preferred to see three separate PRs with their own bechmarks. One referenced in this PR isn't too convincing.
> 
> > 1 - Reuse the iterator when iterating through batches
> 
> This has introduced bug everywhere where `QueryPages` is used with multiple queries, and implementations run them concurrently. (Which is all implementations, except testing ones.) In the end, this optimization may work only at very few places.
> 
> > 2 - Avoid the string copy when parsing the data from the index.
> 
> I would prefer to keep using standard library for conversion in `chunk.go`. Stdlib is typically optimized faster and hand-in-hand with runtime optimizations. By switching to 3rd party library for such basic operations, we're risking not getting those anymore.
> 
> As to the change, personally I'm not a fan of replacing strings with `[]byte`, as it brings complexity to the code (security and concurrency aspects are completely different) That said, I can see how this specific optimization may be useful.
> 
> > 3 - Use an intermediary `IndexQueryEntry` that does not contains the tableName and hashValue strings for lookupEntriesByQueries that doesn't require it.
> 
> This saves 32 bytes per instance. For 1M entries, that's 32 MB. On first sight it looks like very tiny optimization to have a major impact. Can you provide some context where this is useful?

The problem we have in Loki is that compare to cortex, we have many chunks in a small time frame, and so index queries are very expensive because they returns a lot of hits. Now I think I'm going to park this PR because it gave me another idea. I think the problem is more how GetRefChunk works. Most of the time, I need only the first few refs. I'll see if I can refactor this way instead.

I agree that some of those optimization are not yielding enough improvement.