@cyriltovena 2020-12-09
More context 

![image](https://user-images.githubusercontent.com/1053421/101605578-57907d80-39d0-11eb-8e37-148803988256.png)


I've pin down the problem. With actually honor the timerange on the labels and series API. Cortex never did on the chunk store. Still thinking a lot about this.